🔐 Fix login + session, 🚔 police dashboard tips, 🧾 citizen case fetch
- Fixed login flow and stored JWT token in sessionStorage
- Ensured logout properly clears session and redirects to home
- Fetched open cases on citizen dashboard with authorization
- Fetched pending anonymous tips in police dashboard with proper role checks
- Updated anonymousController.js to return 200 on empty results
- Improved tip and case rendering in dashboards
- Minor UI enhancements for clarity and responsiveness
- Created reusable Login component
- Started setup for custom hooks (untracked)
