+++ backend/controllers/userController.js
@@ -41,55 +41,51 @@ export const registerUser = async (req, res) => {
 };
 
 export const loginUser = async (req, res) => {
-    try {
-      const { role } = req.body;
-  
-      if (role === 'anonymous') {
-        return res.status(403).json({ message: 'Anonymous users do not need login.' });
-      }
-  
-      let user;
-  
-      if (role === 'police') {
-        const { name, badgeNumber } = req.body;
-        if (!name || !badgeNumber) return res.status(400).json({ message: 'Name and badge number are required' });
-  
-        user = await User.findOne({ name, badgeNumber });
-      } 
-      
-      else if (role === 'citizen') {
-        const { name, phone } = req.body;
-        if (!name || !phone) return res.status(400).json({ message: 'Name and phone are required' });
-  
-        user = await User.findOne({ name, phone });
-      } 
+  try {
+    const { role } = req.body;
 
-      
-  
-      if (!user) {
-        return res.status(401).json({ message: 'Invalid credentials' });
-      }
-  
-      // Generate Token
-      const token = jwt.sign(
-        {
-          id: user._id,
-          role: user.role,
-          name: user.name,
-          phone: user.phone,
-          badgeNumber: user.badgeNumber
-        },
-        process.env.JWT_SECRET,
-        { expiresIn: '1h' }
-      );
-      
-      
-      console.log("generated token is : ", token);
-
-      res.json({ token, user });
-    } catch (error) {
-      res.status(500).json({ message: 'Login failed' });
+    if (role === 'anonymous') {
+      return res.status(403).json({ message: 'Anonymous users do not need login.' });
     }
+
+    let user;
+
+    if (role === 'police') {
+      const { badgeNumber, password } = req.body;
+      if (!badgeNumber || !password) {
+        return res.status(400).json({ message: 'Badge number and password are required' });
+      }
+      user = await User.findOne({ badgeNumber });
+    }
+
+    else if (role === 'citizen') {
+      const { phone, password } = req.body;
+      if (!phone || !password) {
+        return res.status(400).json({ message: 'Phone and password are required' });
+      }
+      user = await User.findOne({ phone });
+    }
+
+    if (!user) {
+      return res.status(401).json({ message: 'Invalid credentials' });
+    }
+
+    // Generate Token
+    const token = jwt.sign(
+      {
+        id: user._id,
+        role: user.role,
+        name: user.name,
+        phone: user.phone,
+        badgeNumber: user.badgeNumber,
+      },
+      process.env.JWT_SECRET,
+      { expiresIn: '1h' }
+    );
+    res.json({ token, user });
+  } catch (error) {
+    res.status(500).json({ message: 'Login failed' });
+  }
 };
   
 export const findPoliceOfficers = async(req, res) => {
