+++ frontend/app/dashboard/policeDashboard/page.jsx
@@ -127,11 +127,19 @@ const PoliceDashboard = () => {
           </div>
 
           {/* Officer Info on the Right */}
-          <div className="absolute right-4 flex items-center space-x-2">
-            <div className="bg-blue-500/20 p-2 rounded-full">
-              <User className="h-5 w-5 text-blue-300" />
+          <div className="absolute right-4 flex items-center space-x-4">
+            <div className="h-8 w-8 rounded-full bg-gradient-to-r from-cyan-500 to-blue-500 flex items-center justify-center">
+              <User className="h-5 w-5 text-white" />
             </div>
-            <span className="text-blue-300">Officer</span>
+            <button
+              onClick={() => {
+                sessionStorage.clear();
+                window.location.href = "/";
+              }}
+              className="text-sm text-blue-300 hover:text-white underline transition"
+            >
+              Logout
+            </button>
           </div>
         </div>
       </nav>
@@ -240,98 +248,97 @@ const PoliceDashboard = () => {
             <CitizenCasesForPoliceDash />
             {/* Cases List */}
             <div className="bg-slate-900/50 backdrop-blur-md rounded-xl p-6 border border-blue-400/20">
-  <div className="flex items-center justify-between mb-6">
-    <h2 className="text-xl font-semibold text-white flex items-center">
-      <Briefcase className="h-5 w-5 text-blue-400 mr-2" />
-      Active Filed Cases
-    </h2>
-    <div className="flex space-x-2">
-      
-    </div>
-  </div>
+              <div className="flex items-center justify-between mb-6">
+                <h2 className="text-xl font-semibold text-white flex items-center">
+                  <Briefcase className="h-5 w-5 text-blue-400 mr-2" />
+                  Active Filed Cases
+                </h2>
+                <div className="flex space-x-2">
 
-  {/* Horizontal scrolling container */}
-  <div className="overflow-x-auto pb-4" style={{ scrollbarWidth: 'thin', scrollbarColor: '#60a5fa #1e293b' }}>
-    <div className="flex flex-nowrap gap-6 min-w-full">
-      {openCases.length > 0 ? (
-        openCases.map((case_) => (
-          <Link
-            href={`/supportPages/PoliceCaseDetailDisplay/${case_.caseNumber}`}
-            key={case_.caseNumber}
-            className="flex-none w-80"
-          >
-            <div className="bg-white/5 hover:bg-white/10 rounded-2xl p-6 cursor-pointer transition-all shadow-md border-t-4 border-blue-500 h-full">
-              <div className="flex flex-col h-full">
-                <div className="flex justify-between items-start">
-                  <div className="flex items-center space-x-3">
-                    <span className="text-sm text-blue-300 font-mono">
-                      #{case_.caseNumber}
-                    </span>
-                    <span
-                      className={`text-xs px-3 py-1 rounded-full font-semibold ${
-                        case_.priority === "High"
-                          ? "bg-red-500/20 text-red-300"
-                          : case_.priority === "Medium"
-                          ? "bg-yellow-500/20 text-yellow-300"
-                          : "bg-green-500/20 text-green-300"
-                      }`}
-                    >
-                      {case_.priority}
-                    </span>
-                  </div>
-                  <span className="text-xs text-blue-400 whitespace-nowrap bg-blue-900/30 px-3 py-1 rounded-full">
-                    {case_.lastUpdate}
-                  </span>
                 </div>
+              </div>
 
-                <div className="my-4">
-                  <h3 className="text-white text-lg font-semibold mb-2">
-                    {case_.title}
-                  </h3>
-                  <p className="text-blue-200 text-sm line-clamp-3">
-                    {case_.description}
-                  </p>
+              {/* Horizontal scrolling container */}
+              <div className="overflow-x-auto pb-4" style={{ scrollbarWidth: 'thin', scrollbarColor: '#60a5fa #1e293b' }}>
+                <div className="flex flex-nowrap gap-6 min-w-full">
+                  {openCases.length > 0 ? (
+                    openCases.map((case_) => (
+                      <Link
+                        href={`/supportPages/PoliceCaseDetailDisplay/${case_.caseNumber}`}
+                        key={case_.caseNumber}
+                        className="flex-none w-80"
+                      >
+                        <div className="bg-white/5 hover:bg-white/10 rounded-2xl p-6 cursor-pointer transition-all shadow-md border-t-4 border-blue-500 h-full">
+                          <div className="flex flex-col h-full">
+                            <div className="flex justify-between items-start">
+                              <div className="flex items-center space-x-3">
+                                <span className="text-sm text-blue-300 font-mono">
+                                  #{case_.caseNumber}
+                                </span>
+                                <span
+                                  className={`text-xs px-3 py-1 rounded-full font-semibold ${case_.priority === "High"
+                                    ? "bg-red-500/20 text-red-300"
+                                    : case_.priority === "Medium"
+                                      ? "bg-yellow-500/20 text-yellow-300"
+                                      : "bg-green-500/20 text-green-300"
+                                    }`}
+                                >
+                                  {case_.priority}
+                                </span>
+                              </div>
+                              <span className="text-xs text-blue-400 whitespace-nowrap bg-blue-900/30 px-3 py-1 rounded-full">
+                                {case_.lastUpdate}
+                              </span>
+                            </div>
+
+                            <div className="my-4">
+                              <h3 className="text-white text-lg font-semibold mb-2">
+                                {case_.title}
+                              </h3>
+                              <p className="text-blue-200 text-sm line-clamp-3">
+                                {case_.description}
+                              </p>
+                            </div>
+
+                            <div className="mt-auto flex flex-col gap-3 text-sm text-blue-200">
+                              <div className="flex items-center space-x-2">
+                                <MapPin className="h-4 w-4 text-blue-400 flex-shrink-0" />
+                                <span className="truncate">{case_.location}</span>
+                              </div>
+
+                              <div className="flex items-center space-x-2">
+                                <User className="h-4 w-4 text-blue-400 flex-shrink-0" />
+                                <span className="truncate">{case_.assignedTo}</span>
+                              </div>
+                            </div>
+                          </div>
+                        </div>
+                      </Link>
+                    ))
+                  ) : (
+                    <div className="bg-blue-900/20 text-blue-300 p-8 rounded-xl text-center w-full">
+                      <FileX className="h-12 w-12 mx-auto mb-2 text-blue-400 opacity-70" />
+                      <p className="text-white text-center">
+                        No active cases available.
+                      </p>
+                    </div>
+                  )}
                 </div>
+              </div>
 
-                <div className="mt-auto flex flex-col gap-3 text-sm text-blue-200">
-                  <div className="flex items-center space-x-2">
-                    <MapPin className="h-4 w-4 text-blue-400 flex-shrink-0" />
-                    <span className="truncate">{case_.location}</span>
-                  </div>
-
-                  <div className="flex items-center space-x-2">
-                    <User className="h-4 w-4 text-blue-400 flex-shrink-0" />
-                    <span className="truncate">{case_.assignedTo}</span>
-                  </div>
+              {/* Navigation controls */}
+              <div className="mt-4 flex justify-between items-center">
+                <span className="text-blue-300 text-sm">{openCases.length} active cases</span>
+                <div className="flex space-x-3">
+                  <button className="bg-blue-600/30 hover:bg-blue-600/40 text-blue-300 p-2 rounded-full transition-colors">
+                    <ChevronLeft className="h-5 w-5" />
+                  </button>
+                  <button className="bg-blue-600/30 hover:bg-blue-600/40 text-blue-300 p-2 rounded-full transition-colors">
+                    <ChevronRight className="h-5 w-5" />
+                  </button>
                 </div>
               </div>
             </div>
-          </Link>
-        ))
-      ) : (
-        <div className="bg-blue-900/20 text-blue-300 p-8 rounded-xl text-center w-full">
-          <FileX className="h-12 w-12 mx-auto mb-2 text-blue-400 opacity-70" />
-          <p className="text-white text-center">
-            No active cases available.
-          </p>
-        </div>
-      )}
-    </div>
-  </div>
-  
-  {/* Navigation controls */}
-  <div className="mt-4 flex justify-between items-center">
-    <span className="text-blue-300 text-sm">{openCases.length} active cases</span>
-    <div className="flex space-x-3">
-      <button className="bg-blue-600/30 hover:bg-blue-600/40 text-blue-300 p-2 rounded-full transition-colors">
-        <ChevronLeft className="h-5 w-5" />
-      </button>
-      <button className="bg-blue-600/30 hover:bg-blue-600/40 text-blue-300 p-2 rounded-full transition-colors">
-        <ChevronRight className="h-5 w-5" />
-      </button>
-    </div>
-  </div>
-</div>
 
             {/* New Section: Citizen Complaints */}
           </div>
@@ -343,36 +350,41 @@ const PoliceDashboard = () => {
                 Active Anonymous Tips
               </h2>
               <div>
-                {anonymousTips.map((tip) => (
-                  <Link
-                    href={`/supportPages/AnonymousTipDetailDisplay/${tip._id}`}
-                    key={tip._id}
-                  >
-                    <div className="bg-white/5 rounded-xl p-4 hover:bg-white/10 transition-all mb-4">
-                      <div className="flex items-center justify-between">
-                        <span className="text-sm text-blue-300">
-                          <p>
-                            Coordinates:{" "}
-                            {tip.location.coordinates[0]}, {" "},
-                            {tip.location.coordinates[1]}
-                          </p>
-                        </span>
-                        <span className="text-xs px-3 py-1 rounded-full bg-red-600 text-blue-200">
-                          {tip.status}
+                {anonymousTips.length > 0 ? (
+                  anonymousTips.map((tip) => (
+                    <Link
+                      href={`/supportPages/AnonymousTipDetailDisplay/${tip._id}`}
+                      key={tip._id}
+                    >
+                      <div className="bg-white/5 rounded-xl p-4 hover:bg-white/10 transition-all mb-4">
+                        <div className="flex items-center justify-between">
+                          <span className="text-sm text-blue-300">
+                            <p>
+                              Coordinates:{" "}
+                              {tip.location.coordinates[0]}, {" "},
+                              {tip.location.coordinates[1]}
+                            </p>
+                          </span>
+                          <span className="text-xs px-3 py-1 rounded-full bg-red-600 text-blue-200">
+                            {tip.status}
+                          </span>
+                        </div>
+                        <p className="text-white text-sm mt-2">
+                          {tip.description}
+                        </p>
+                        <span className="text-sm text-blue-300 mt-2 block">
+                          {findTimeDifference(tip.submitted_at)} hour(s) ago
                         </span>
                       </div>
-                      <p className="text-white text-sm mt-2">
-                        {tip.description}
-                      </p>
-                      <span className="text-sm text-blue-300 mt-2 block">
-                        {findTimeDifference(tip.submitted_at)} hour(s) ago
-                      </span>
-                    </div>
-                  </Link>
-                ))}
+                    </Link>
+                  ))
+                ) : (
+                  <div className="text-blue-300 text-sm bg-white/5 p-4 rounded-lg">
+                    No anonymous tips found.
+                  </div>
+                )}
               </div>
             </div>
-
             {/* Patrol Units */}
             <PatrolUnits />
           </div>
